requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
tensorflow-cpu==2.10.1
numpy==1.23.5    
Pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.8.4
scipy==1.10.1